small fix for Windows build

+++ vlib/clipboard/clipboard_windows.c.v
@@ -127,7 +127,7 @@ fn (mut cb Clipboard) free() {
 fn to_wide(text string) &C.HGLOBAL {
 	len_required := C.MultiByteToWideChar(C.CP_UTF8, C.MB_ERR_INVALID_CHARS, text.str, text.len +
 		1, C.NULL, 0)
-    buf := C.GlobalAlloc(C.GMEM_MOVEABLE, sizeof(u16) * len_required)
+    buf := C.GlobalAlloc(C.GMEM_MOVEABLE, i64(sizeof(u16)) * len_required)
     if buf != C.HGLOBAL(C.NULL) {
         mut locked := &u16(C.GlobalLock(buf))
         C.MultiByteToWideChar(C.CP_UTF8, C.MB_ERR_INVALID_CHARS, text.str, text.len + 1, locked, len_required)
